src/tradingbot/preprocessor/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
frozendict==2.3.8
greenlet==3.0.3
html5lib==1.1
idna==3.4
lxml==4.9.3
multitasking==0.0.11
numpy==1.25.1
pandas==2.0.3
peewee==3.17.6
platformdirs==4.2.2
psycopg==3.2.1
psycopg-pool==3.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.31
ta==0.10.2
tqdm==4.65.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
yfinance==0.2.41
